BIM: move SketchArch hooks to ArchComponent parent class to avoid duplication
