+++ src/Mod/BIM/ArchWindow.py
@@ -116,23 +116,6 @@ class _Window(ArchComponent.Component):
         # Add features in the SketchArch External Add-on
         self.addSketchArchFeatures(obj)
 
-    def addSketchArchFeatures(self, obj, linkObj=None, mode=None):
-        """
-        To add features in the SketchArch External Add-on  (https://github.com/paullee0/FreeCAD_SketchArch)
-        -  import ArchSketchObject module, and
-        -  set properties that are common to ArchObjects (including Links) and ArchSketch
-           to support the additional features
-
-        To install SketchArch External Add-on, see https://github.com/paullee0/FreeCAD_SketchArch#iv-install
-        """
-
-        try:
-            import ArchSketchObject
-
-            ArchSketchObject.ArchSketch.setPropertiesLinkCommon(self, obj, linkObj, mode)
-        except:
-            pass
-
     def setProperties(self, obj, mode=None):
 
         lp = obj.PropertiesList
@@ -677,38 +660,6 @@ class _Window(ArchComponent.Component):
             obj.Area = obj.Width.Value * obj.Height.Value
 
         self.executeSketchArchFeatures(obj)
-
-    def executeSketchArchFeatures(self, obj, linkObj=None, index=None, linkElement=None):
-        """
-        To execute features in the SketchArch External Add-on  (https://github.com/paullee0/FreeCAD_SketchArch)
-        -  import ArchSketchObject module, and
-        -  execute features that are common to ArchObjects (including Links) and ArchSketch
-
-        To install SketchArch External Add-on, see https://github.com/paullee0/FreeCAD_SketchArch#iv-install
-        """
-
-        # To execute features in SketchArch External Add-on
-        try:
-            import ArchSketchObject  # Why needed ? Should have try: addSketchArchFeatures() before !  Need 'per method' ?
-
-            # Execute SketchArch Feature - Intuitive Automatic Placement for Arch Windows/Doors, Equipment etc.
-            # see https://forum.freecad.org/viewtopic.php?f=23&t=50802
-            ArchSketchObject.updateAttachmentOffset(obj, linkObj)
-        except:
-            pass
-
-    def appLinkExecute(self, obj, linkObj, index, linkElement):
-        """
-        Default Link Execute method() -
-        See https://forum.freecad.org/viewtopic.php?f=22&t=42184&start=10#p361124
-        @realthunder added support to Links to run Linked Scripted Object's methods()
-        """
-
-        # Add features in the SketchArch External Add-on
-        self.addSketchArchFeatures(obj, linkObj)
-
-        # Execute features in the SketchArch External Add-on
-        self.executeSketchArchFeatures(obj, linkObj)
 
     def getSubFace(self):
         "returns a subface for creation of subvolume for cutting in a base object"
